refactor _link method
This was a normalizing refactor. There was some code in both of these two
styles:

  * ask-for-forgiveness: try to create the link first, then handle any error
    conditions that occur
  * look-before-leaping: check if the conditions are good to create the link,
    then try to create it

Even though the ask-for-forgivenmess is generally more robust and helps prefent
certain race conditions, I went with look-before-leaping.... for not much of a
reason. At least the code is consistent now, and I think that's a net
improvement.